requirements.txt
python-telegram-bot==20.0a2
pandas==1.4.2
numpy==1.22.3
